Update to genesPathway function
Update to genesPathway function in module_new.R file. Correction to argument seurat_obj that incorrectly referenced the `master_seurat` object that is not always loaded into scripts (as it is not loaded in the module_new.R file that only contains functions). Clarified this argument to be an empty list. Corrected all figure scripts using the function to specify the Seurat object referenced, note that I may have missed some instances in other scripts, which will require trial and error later.